requirements.txt
alembic==1.0.0
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
Babel==2.7.0
bcrypt==3.1.7
billiard==3.3.0.23
blinker==1.4
celery==3.1.26.post2
certifi==2017.7.27.1
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
concurrent-log-handler==0.9.12
configparser==3.7.4
constantly==15.1.0
crypto==1.4.1
cryptography==2.7
defusedxml==0.5.0
Django==2.0.3
django-adminlte-ui==1.4.0
django-celery==3.2.2
django-cors-headers==3.1.0
django-filter==2.0.0
django-filters==0.2.1
django-pure-pagination==0.3.0
django-redis==4.10.0
django-treebeard==4.3
dwebsocket==0.4.2
elasticsearch==6.3.1
emoji==0.5.2
eventlet==0.22.1
filetype==1.0.5
Flask==1.0.2
Flask-Celery==2.4.3
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
flower==0.9.2
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
gunicorn==19.7.1
har2case==0.3.1
hiredis==1.0.0
honcho==1.0.1
HttpRunner==1.5.8
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
kafka-python==1.4.3
kombu==3.0.37
ldap3==2.5.1
locust==0.0
locustio==0.12.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.15.4
msgpack-python==0.5.6
mysqlclient==1.4.2
Naked==0.1.31
oauthlib==2.0.6
olefile==0.44
paramiko==2.4.1
pbkdf2==1.3
pika==0.12.0
Pillow==5.2.0
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.9.0
pycryptosat==0.2.0
PyHamcrest==1.9.0
PyJWT==1.5.3
pymssql==2.1.4
PyMySQL==0.9.2
PyNaCl==1.3.0
pyotp==2.2.6
pypiwin32==223
pyrsistent==0.15.3
python-dateutil==2.8.0
python-editor==1.0.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
pywin32==224
PyYAML==3.12
pyzmq==18.1.0
raven==6.9.0
redis==3.2.1
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rsa==4.0
selenium==3.141.0
shellescape==3.4.1
simplejson==3.16.0
six==1.11.0
social-auth-core==1.4.0
sorl-thumbnail==12.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
tornado==5.1.1
txaio==18.8.1
ujson==1.35
urllib3==1.22
vine==1.3.0
virtualenv==16.7.7
Werkzeug==0.16.0
zope.interface==4.6.0
